+++ src/main/java/org/opensearch/security/privileges/PrivilegesEvaluator.java
@@ -26,6 +26,7 @@
 
 package org.opensearch.security.privileges;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -93,6 +94,7 @@ import org.opensearch.security.privileges.actionlevel.RoleBasedActionPrivileges;
 import org.opensearch.security.privileges.actionlevel.SubjectBasedActionPrivileges;
 import org.opensearch.security.resolver.IndexResolverReplacer;
 import org.opensearch.security.resolver.IndexResolverReplacer.Resolved;
+import org.opensearch.security.resources.ResourceSharingIndexHandler;
 import org.opensearch.security.securityconf.ConfigModel;
 import org.opensearch.security.securityconf.DynamicConfigFactory;
 import org.opensearch.security.securityconf.DynamicConfigModel;
@@ -154,6 +156,7 @@ public class PrivilegesEvaluator {
     private final ProtectedIndexAccessEvaluator protectedIndexAccessEvaluator;
     private final TermsAggregationEvaluator termsAggregationEvaluator;
     private final PitPrivilegesEvaluator pitPrivilegesEvaluator;
+    private final ResourceAccessEvaluator resourceAccessEvaluator;
     private DynamicConfigModel dcm;
     private final Settings settings;
     private final AtomicReference<RoleBasedActionPrivileges> actionPrivileges = new AtomicReference<>();
@@ -179,7 +182,9 @@ public class PrivilegesEvaluator {
         final Settings settings,
         final PrivilegesInterceptor privilegesInterceptor,
         final ClusterInfoHolder clusterInfoHolder,
-        final IndexResolverReplacer irr
+        final IndexResolverReplacer irr,
+        Set<String> resourceIndices,
+        final ResourceSharingIndexHandler resourceSharingIndexHandler
     ) {
 
         super();
@@ -228,6 +233,7 @@ public class PrivilegesEvaluator {
                 }
             });
         }
+        this.resourceAccessEvaluator = new ResourceAccessEvaluator(resourceIndices, threadPool, resourceSharingIndexHandler);
 
     }
 
@@ -430,11 +436,13 @@ public class PrivilegesEvaluator {
         }
 
         // check snapshot/restore requests
+        // NOTE: Has to go first as restore request could be for protected and/or system indices and the request may
+        // fail with 403 if system index or protected index evaluators are triggered first
         if (snapshotRestoreEvaluator.evaluate(request, task, action0, clusterInfoHolder, presponse).isComplete()) {
             return presponse;
         }
 
-        // Security index access
+        // System index access
         if (systemIndexAccessEvaluator.evaluate(request, task, action0, requestedResolved, presponse, context, actionPrivileges, user)
             .isComplete()) {
             return presponse;
@@ -443,6 +451,15 @@ public class PrivilegesEvaluator {
         // Protected index access
         if (protectedIndexAccessEvaluator.evaluate(request, task, action0, requestedResolved, presponse, mappedRoles).isComplete()) {
             return presponse;
+        }
+
+        // Protected Resources access
+        try {
+            if (resourceAccessEvaluator.evaluate(request, action0, context, presponse).isComplete()) {
+                return presponse;
+            }
+        } catch (IOException e) {
+            // Do nothing
         }
 
         // check access for point in time requests